+++ core/src/base-ingredient.ts
@@ -1,19 +1,19 @@
-import { IBakeConfig, IIngredient, IBakeRegion, IBakeAuthentication} from "./bake-interfaces";
-import {Logger} from './logger'
-import {DeploymentContext} from './deployment-context'
+import { IBakeConfig, IIngredient, IBakeRegion, IBakeAuthentication } from "./bake-interfaces";
+import { Logger } from './logger'
+import { DeploymentContext } from './deployment-context'
 import * as colors from 'colors'
 
 export class BaseIngredient {
 
     constructor(name: string, ingredient: IIngredient, ctx: DeploymentContext) {
 
-        this._ctx = new DeploymentContext(ctx.AuthToken, ctx.Package, ctx.Region, ctx.Logger,ingredient)
+        this._ctx = new DeploymentContext(ctx.AuthToken, ctx.Package, ctx.Region, ctx.Logger, ingredient, ctx.CustomAuthToken)
         this._name = name
         this._ingredient = ingredient
         this._logger = new Logger(ctx.Logger.getPre().concat(name), ctx.Environment.logLevel)
 
-        ingredient.properties.source.valueAsync(ctx).then(source=>{
-            this._logger.log('adding ingredient type[' + colors.cyan(ingredient.properties.type) +'] source[' + colors.cyan(source) + ']')
+        ingredient.properties.source.valueAsync(ctx).then(source => {
+            this._logger.log('adding ingredient type[' + colors.cyan(ingredient.properties.type) + '] source[' + colors.cyan(source) + ']')
         })
     }
 
@@ -26,6 +26,6 @@ export class BaseIngredient {
     }
 
     public async Auth(auth: IBakeAuthentication): Promise<string | null> {
-        return null;       
+        return null;
     }
 }