+++ ingredient/ingredient-property-service/src/models/property.ts
@@ -1,0 +1,7 @@
+import { IdentifierBase } from ".";
+
+export class PropertyIdentifier extends IdentifierBase {
+    constructor(name: string, identifier: string) {
+        super(name, identifier);
+    }
+}