+++ ingredient/ingredient-property-service/src/models/searchOperator.ts
@@ -1,0 +1,5 @@
+export enum SearchOperator {
+    None = 0,
+    Contains = 1,
+    Equals = 2
+}