+++ ingredient/ingredient-property-service/src/propertyTypes.ts
@@ -1,0 +1,6 @@
+export enum PropertyType {
+    Property = 'property',
+    Secret = 'secret',
+    EncryptionKey = 'encryptionkey',
+    Certificate = 'certificate'
+}