+++ ingredient/ingredient-property-service/src/validations/index.ts
@@ -1,0 +1,2 @@
+export { PropertyConfigurationValidator } from './propertyValidations'
+export { SecretConfigurationValidator } from './secretValidations'